pyproject.toml
[build-system]
requires = [
    "setuptools>=45", "setuptools_scm[toml]>=6.2"
]
build-backend = "setuptools.build_meta"

[project]
name = "specterext_taxtheft"

authors = [
  { name="k9ert"},
]
description = "A way to deal with the fallout of taxtheft"
urls = { Homepage = "https://github.com/cryptoadvance/specterext-stacktrack" }
readme = "README.md"

requires-python = ">=3.9"

dependencies = [
    "ccxt",
    "pandas"
]

license = {file = "LICENSE"}

classifiers =[
    'Programming Language :: Python :: 3',
    'License :: OSI Approved :: MIT License',
    'Operating System :: OS Independent',
]
dynamic=["version"]

[tool.setuptools_scm]
write_to = "src/k9ert/specterext/taxtheft/_version.py"

[project.optional-dependencies]
test = [
  "cryptoadvance.specter>=1.14.5",
  "pytest >=7.1.3",
  "pytest-cov[all]"
]